Index existing docs on index creation
