+++ db/collection_index.go
@@ -13,12 +13,18 @@ package db
 import (
 	"context"
 	"encoding/json"
+	"fmt"
+	"strconv"
+	"strings"
 
+	ds "github.com/ipfs/go-datastore"
 	"github.com/ipfs/go-datastore/query"
 
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/core"
 	"github.com/sourcenetwork/defradb/datastore"
+	"github.com/sourcenetwork/defradb/db/base"
+	"github.com/sourcenetwork/defradb/db/fetcher"
 )
 
 // collectionIndexDescription describes an index on a collection.
@@ -190,4 +196,325 @@ func (c *collection) updateIndex(
 		}
 	}
 	return nil
+}
+
+func (c *collection) CreateIndex(
+	ctx context.Context,
+	desc client.IndexDescription,
+) (client.IndexDescription, error) {
+	txn, err := c.getTxn(ctx, false)
+	if err != nil {
+		return client.IndexDescription{}, err
+	}
+
+	index, err := c.createIndex(ctx, txn, desc)
+	if err != nil {
+		return client.IndexDescription{}, err
+	}
+	if c.isIndexCached {
+		c.indexes = append(c.indexes, index)
+	}
+	err = c.indexExistingDocs(ctx, txn, index)
+	if err != nil {
+		return client.IndexDescription{}, err
+	}
+	return index.Description(), nil
+}
+
+func (c *collection) newFetcher() fetcher.Fetcher {
+	if c.fetcherFactory != nil {
+		return c.fetcherFactory()
+	} else {
+		return new(fetcher.DocumentFetcher)
+	}
+}
+func (c *collection) indexExistingDocs(
+	ctx context.Context,
+	txn datastore.Txn,
+	index CollectionIndex,
+) error {
+	df := c.newFetcher()
+
+	fields := make([]*client.FieldDescription, 0, 1)
+	for _, field := range index.Description().Fields {
+		for i := range c.desc.Schema.Fields {
+			colField := &c.desc.Schema.Fields[i]
+			if field.Name == colField.Name {
+				fields = append(fields, colField)
+				break
+			}
+		}
+	}
+
+	err := df.Init(&c.desc, fields, false, false)
+	if err != nil {
+		_ = df.Close()
+		return err
+	}
+	start := base.MakeCollectionKey(c.desc)
+	spans := core.NewSpans(core.NewSpan(start, start.PrefixEnd()))
+
+	err = df.Start(ctx, txn, spans)
+	if err != nil {
+		_ = df.Close()
+		return err
+	}
+
+	var doc *client.Document
+	for {
+		doc, err = df.FetchNextDecoded(ctx)
+		if err != nil {
+			_ = df.Close()
+			return err
+		}
+		if doc == nil {
+			break
+		}
+		err = index.Save(ctx, txn, doc)
+		if err != nil {
+			return err
+		}
+	}
+
+	return df.Close()
+}
+
+func (c *collection) DropIndex(ctx context.Context, indexName string) error {
+	key := core.NewCollectionIndexKey(c.Name(), indexName)
+
+	txn, err := c.getTxn(ctx, false)
+	if err != nil {
+		return err
+	}
+	_, err = c.getIndexes(ctx, txn)
+	if err != nil {
+		return err
+	}
+	for i := range c.indexes {
+		if c.indexes[i].Name() == indexName {
+			err = c.indexes[i].RemoveAll(ctx, txn)
+			if err != nil {
+				return err
+			}
+			c.indexes = append(c.indexes[:i], c.indexes[i+1:]...)
+			break
+		}
+	}
+	err = txn.Systemstore().Delete(ctx, key.ToDS())
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (c *collection) dropAllIndexes(ctx context.Context, txn datastore.Txn) error {
+	prefix := core.NewCollectionIndexKey(c.Name(), "")
+
+	err := iteratePrefixKeys(ctx, prefix.ToString(), txn.Systemstore(),
+		func(ctx context.Context, key ds.Key) error {
+			return txn.Systemstore().Delete(ctx, key)
+		})
+
+	return err
+}
+
+func (c *collection) getIndexes(ctx context.Context, txn datastore.Txn) ([]CollectionIndex, error) {
+	if c.isIndexCached {
+		return c.indexes, nil
+	}
+
+	prefix := core.NewCollectionIndexKey(c.Name(), "")
+	indexes, err := deserializePrefix[client.IndexDescription](ctx, prefix.ToString(), txn.Systemstore())
+	if err != nil {
+		return nil, err
+	}
+	colIndexes := make([]CollectionIndex, 0, len(indexes))
+	for _, index := range indexes {
+		colIndexes = append(colIndexes, NewCollectionIndex(c, index))
+	}
+
+	c.indexes = colIndexes
+	c.isIndexCached = true
+	return colIndexes, nil
+}
+
+func (c *collection) GetIndexes(ctx context.Context) ([]client.IndexDescription, error) {
+	txn, err := c.getTxn(ctx, true)
+	if err != nil {
+		return nil, err
+	}
+	indexes, err := c.getIndexes(ctx, txn)
+	if err != nil {
+		return nil, err
+	}
+	indexDescriptions := make([]client.IndexDescription, 0, len(indexes))
+	for _, index := range indexes {
+		indexDescriptions = append(indexDescriptions, index.Description())
+	}
+
+	return indexDescriptions, nil
+}
+
+func (c *collection) createIndex(
+	ctx context.Context,
+	txn datastore.Txn,
+	desc client.IndexDescription,
+) (CollectionIndex, error) {
+	err := validateIndexDescription(desc)
+	if err != nil {
+		return nil, err
+	}
+
+	err = c.checkExistingFields(ctx, desc.Fields)
+	if err != nil {
+		return nil, err
+	}
+
+	indexKey, err := c.processIndexName(ctx, txn, &desc)
+	if err != nil {
+		return nil, err
+	}
+
+	colSeq, err := c.db.getSequence(ctx, txn, fmt.Sprintf("%s/%d", core.COLLECTION_INDEX, c.ID()))
+	if err != nil {
+		return nil, err
+	}
+	colID, err := colSeq.next(ctx, txn)
+	if err != nil {
+		return nil, err
+	}
+	desc.ID = uint32(colID)
+
+	buf, err := json.Marshal(desc)
+	if err != nil {
+		return nil, err
+	}
+
+	err = txn.Systemstore().Put(ctx, indexKey.ToDS(), buf)
+	if err != nil {
+		return nil, err
+	}
+	colIndex := NewCollectionIndex(c, desc)
+	return colIndex, nil
+}
+
+func (c *collection) checkExistingFields(
+	ctx context.Context,
+	fields []client.IndexedFieldDescription,
+) error {
+	collectionFields := c.Description().Schema.Fields
+	for _, field := range fields {
+		found := false
+		fieldLower := strings.ToLower(field.Name)
+		for _, colField := range collectionFields {
+			if fieldLower == strings.ToLower(colField.Name) {
+				found = true
+				break
+			}
+		}
+		if !found {
+			return NewErrNonExistingFieldForIndex(field.Name)
+		}
+	}
+	return nil
+}
+
+func (c *collection) processIndexName(
+	ctx context.Context,
+	txn datastore.Txn,
+	desc *client.IndexDescription,
+) (core.CollectionIndexKey, error) {
+	var indexKey core.CollectionIndexKey
+	if desc.Name == "" {
+		nameIncrement := 1
+		for {
+			desc.Name = generateIndexName(c, desc.Fields, nameIncrement)
+			indexKey = core.NewCollectionIndexKey(c.Name(), desc.Name)
+			exists, err := txn.Systemstore().Has(ctx, indexKey.ToDS())
+			if err != nil {
+				return core.CollectionIndexKey{}, err
+			}
+			if !exists {
+				break
+			}
+			nameIncrement++
+		}
+	} else {
+		indexKey = core.NewCollectionIndexKey(c.Name(), desc.Name)
+		exists, err := txn.Systemstore().Has(ctx, indexKey.ToDS())
+		if err != nil {
+			return core.CollectionIndexKey{}, err
+		}
+		if exists {
+			return core.CollectionIndexKey{}, ErrIndexWithNameAlreadyExists
+		}
+	}
+	return indexKey, nil
+}
+
+func validateIndexDescription(desc client.IndexDescription) error {
+	if desc.ID != 0 {
+		return NewErrNonZeroIndexIDProvided(desc.ID)
+	}
+	if len(desc.Fields) == 0 {
+		return ErrIndexMissingFields
+	}
+	if len(desc.Fields) == 1 && desc.Fields[0].Direction == client.Descending {
+		return ErrIndexSingleFieldWrongDirection
+	}
+	for i := range desc.Fields {
+		if desc.Fields[i].Name == "" {
+			return ErrIndexFieldMissingName
+		}
+		if desc.Fields[i].Direction == "" {
+			desc.Fields[i].Direction = client.Ascending
+		}
+	}
+	return nil
+}
+
+func generateIndexName(col client.Collection, fields []client.IndexedFieldDescription, inc int) string {
+	sb := strings.Builder{}
+	direction := "ASC"
+	//if fields[0].Direction == client.Descending {
+	//direction = "DESC"
+	//}
+	sb.WriteString(col.Name())
+	sb.WriteByte('_')
+	sb.WriteString(fields[0].Name)
+	sb.WriteByte('_')
+	sb.WriteString(direction)
+	if inc > 1 {
+		sb.WriteByte('_')
+		sb.WriteString(strconv.Itoa(inc))
+	}
+	return sb.String()
+}
+
+func deserializePrefix[T any](ctx context.Context, prefix string, storage ds.Read) ([]T, error) {
+	q, err := storage.Query(ctx, query.Query{Prefix: prefix})
+	if err != nil {
+		return nil, NewErrFailedToCreateCollectionQuery(err)
+	}
+	defer func() {
+		if err := q.Close(); err != nil {
+			log.ErrorE(ctx, "Failed to close collection query", err)
+		}
+	}()
+
+	elements := make([]T, 0)
+	for res := range q.Next() {
+		if res.Error != nil {
+			return nil, res.Error
+		}
+
+		var element T
+		err = json.Unmarshal(res.Value, &element)
+		if err != nil {
+			return nil, NewErrInvalidStoredIndex(err)
+		}
+		elements = append(elements, element)
+	}
+	return elements, nil
 }
