+++ db/indexed_docs_test.go
@@ -252,7 +252,6 @@ func TestNonUnique_IfDocIsAdded_ShouldBeIndexed(t *testing.T) {
 
 	doc := f.newUserDoc("John", 21)
 	f.saveDocToCollection(doc, f.users)
-	//f.commitTxn()
 
 	key := newIndexKeyBuilder(f).Col(usersColName).Field(usersNameFieldName).Doc(doc).Build()
 
@@ -505,6 +504,121 @@ func TestNonUnique_IfIndexedFieldIsNil_StoreItAsNil(t *testing.T) {
 	data, err := f.txn.Datastore().Get(f.ctx, key.ToDS())
 	require.NoError(t, err)
 	assert.Len(t, data, 0)
+}
+
+func TestNonUniqueCreate_ShouldIndexExistingDocs(t *testing.T) {
+	f := newIndexTestFixture(t)
+
+	doc1 := f.newUserDoc("John", 21)
+	f.saveDocToCollection(doc1, f.users)
+	doc2 := f.newUserDoc("Islam", 18)
+	f.saveDocToCollection(doc2, f.users)
+
+	f.createUserCollectionIndexOnName()
+
+	key1 := newIndexKeyBuilder(f).Col(usersColName).Field(usersNameFieldName).Doc(doc1).Build()
+	key2 := newIndexKeyBuilder(f).Col(usersColName).Field(usersNameFieldName).Doc(doc2).Build()
+
+	data, err := f.txn.Datastore().Get(f.ctx, key1.ToDS())
+	require.NoError(t, err)
+	assert.Len(t, data, 0)
+	data, err = f.txn.Datastore().Get(f.ctx, key2.ToDS())
+	require.NoError(t, err)
+	assert.Len(t, data, 0)
+}
+
+func TestNonUniqueCreate_IfUponIndexingExistingDocsFetcherFails_ReturnError(t *testing.T) {
+	testError := errors.New("test error")
+
+	cases := []struct {
+		Name           string
+		PrepareFetcher func() fetcher.Fetcher
+	}{
+		{
+			Name: "Fails to init",
+			PrepareFetcher: func() fetcher.Fetcher {
+				f := fetcherMocks.NewStubbedFetcher(t)
+				f.EXPECT().Init(mock.Anything, mock.Anything, mock.Anything, mock.Anything).Unset()
+				f.EXPECT().Init(mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(testError)
+				f.EXPECT().Close().Unset()
+				f.EXPECT().Close().Return(nil)
+				return f
+			},
+		},
+		{
+			Name: "Fails to start",
+			PrepareFetcher: func() fetcher.Fetcher {
+				f := fetcherMocks.NewStubbedFetcher(t)
+				f.EXPECT().Start(mock.Anything, mock.Anything, mock.Anything).Unset()
+				f.EXPECT().Start(mock.Anything, mock.Anything, mock.Anything).Return(testError)
+				f.EXPECT().Close().Unset()
+				f.EXPECT().Close().Return(nil)
+				return f
+			},
+		},
+		{
+			Name: "Fails to fetch next decoded",
+			PrepareFetcher: func() fetcher.Fetcher {
+				f := fetcherMocks.NewStubbedFetcher(t)
+				f.EXPECT().FetchNextDecoded(mock.Anything).Unset()
+				f.EXPECT().FetchNextDecoded(mock.Anything).Return(nil, testError)
+				f.EXPECT().Close().Unset()
+				f.EXPECT().Close().Return(nil)
+				return f
+			},
+		},
+		{
+			Name: "Fails to close",
+			PrepareFetcher: func() fetcher.Fetcher {
+				f := fetcherMocks.NewStubbedFetcher(t)
+				f.EXPECT().FetchNextDecoded(mock.Anything).Unset()
+				f.EXPECT().FetchNextDecoded(mock.Anything).Return(nil, nil)
+				f.EXPECT().Close().Unset()
+				f.EXPECT().Close().Return(testError)
+				return f
+			},
+		},
+	}
+
+	for _, tc := range cases {
+		f := newIndexTestFixture(t)
+
+		doc := f.newUserDoc("John", 21)
+		f.saveDocToCollection(doc, f.users)
+
+		f.users.fetcherFactory = tc.PrepareFetcher
+		key := newIndexKeyBuilder(f).Col(usersColName).Field(usersNameFieldName).Doc(doc).Build()
+
+		_, err := f.users.CreateIndex(f.ctx, getUsersIndexDescOnName())
+		require.ErrorIs(t, err, testError, tc.Name)
+
+		_, err = f.txn.Datastore().Get(f.ctx, key.ToDS())
+		require.Error(t, err, tc.Name)
+	}
+}
+
+func TestNonUniqueCreate_IfDatastoreFailsToStoreIndex_ReturnError(t *testing.T) {
+	f := newIndexTestFixture(t)
+
+	doc := f.newUserDoc("John", 21)
+	f.saveDocToCollection(doc, f.users)
+
+	testErr := errors.New("test error")
+
+	f.users.fetcherFactory = func() fetcher.Fetcher {
+		return fetcherMocks.NewStubbedFetcher(t)
+	}
+
+	mockedTxn := f.mockTxn()
+	mockedTxn.MockDatastore = mocks.NewDSReaderWriter(t)
+	mockedTxn.MockDatastore.EXPECT().Put(mock.Anything, mock.Anything, mock.Anything).Unset()
+	mockedTxn.MockDatastore.EXPECT().Put(mock.Anything, mock.Anything, mock.Anything).
+		Return(testErr)
+	mockedTxn.EXPECT().Datastore().Unset()
+	mockedTxn.EXPECT().Datastore().Return(mockedTxn.MockDatastore)
+
+	_, err := f.users.WithTxn(mockedTxn).CreateIndex(f.ctx, getUsersIndexDescOnName())
+	require.ErrorIs(f.t, err, testErr)
 }
 
 func TestNonUniqueDrop_ShouldDeleteStoredIndexedFields(t *testing.T) {
